Search functionality / Tweaks :wrench:
