+++ package.json
@@ -6,18 +6,26 @@
     "build": "rollup -c",
     "dev": "rollup -c -w",
     "start": "sirv public --no-clear",
-    "compile": "rm -f -R node_modules && rm -f ./public/bundle.zip && npm ci && npm run build && zip -r -x './public/.env' -X bundle.zip ./public/*"
+    "compile": "rm -f -R node_modules && rm -f ./public/bundle.zip && npm ci && npm run build && zip -r -x './public/.env' -X bundle.zip ./public/*",
+    "check": "svelte-check --tsconfig ./tsconfig.json"
   },
   "devDependencies": {
     "@rollup/plugin-commonjs": "^17.0.0",
     "@rollup/plugin-html": "^0.2.4",
     "@rollup/plugin-node-resolve": "^11.0.0",
+    "@rollup/plugin-typescript": "^8.0.0",
+    "@tsconfig/svelte": "^2.0.1",
     "rollup": "^2.3.4",
     "rollup-plugin-css-only": "^3.1.0",
     "rollup-plugin-livereload": "^2.0.0",
+    "rollup-plugin-string": "^3.0.0",
     "rollup-plugin-svelte": "^7.0.0",
     "rollup-plugin-terser": "^7.0.0",
-    "svelte": "^3.46.4"
+    "svelte": "^3.46.4",
+    "svelte-check": "^2.0.0",
+    "svelte-preprocess": "^4.0.0",
+    "tslib": "^2.0.0",
+    "typescript": "^4.0.0"
   },
   "dependencies": {
     "ace-builds": "^1.4.14",
