+++ src/utils/storage/fixtures/getDatabasesNames.js
@@ -1,0 +1,3 @@
+(async function () {
+  return (window.__INDEXED_DB_EDIT__.tables = await indexedDB.databases());
+})();
