+++ src/utils/storage/fixtures/getFromCache.js
@@ -1,0 +1,3 @@
+(function () {
+  return window.__INDEXED_DB_EDIT__["${key}"];
+})();
