+++ src/utils/storage/fixtures/init.js
@@ -1,0 +1,1 @@
+window.__INDEXED_DB_EDIT__ = {};
